Homework1/requirements.txt
Jinja2
MarkupSafe
Pillow~=9.3.0
PyWavelets
Pygments
appnope
astroid
bleach
cycler
decorator
entrypoints
html5lib
ipykernel
ipython
ipython-genutils
ipywidgets
isort
jedi
jsonschema
jupyter
jupyter-client
jupyter-console
jupyter-core
lazy-object-proxy
matplotlib
mccabe
mistune
nbconvert
nbformat
networkx
notebook
numpy~=1.21.4
olefile
pandocfilters
pexpect
pickleshare
prompt-toolkit
ptyprocess
pylint
pyparsing
python-dateutil
pytz
pyzmq
qtconsole
scikit-image~=0.20.0
scipy
simplegeneric
six
terminado
testpath
timeout-decorator
tornado
wcwidth
webencodings
widgetsnbextension
wrapt
